Separate type codes into files in type directory
