+++ app/types/editable-dynamic-component.ts
@@ -1,0 +1,5 @@
+export interface EditableDynamicComponent {
+  id: number;
+  componentName: keyof JSX.IntrinsicElements;
+  innerText?: string;
+}
